+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Added Poisonous Potato Crate.
 - Added Bread Bag.
+- Added Pumpkin Slice Crate (for Farmer's Delight).
 - Potato Crates can now be smelted in a furnace to get 9 potato items.
 
 ***